put the interpreter in a separate directory
