+++ src/Makefile
@@ -7,11 +7,6 @@ OBJNAMES := \
 	ast.o \
 	compile_time_environment.o \
 	desugar.o \
-	environment.o \
-	error.o \
-	eval.o \
-	execute.o \
-	garbage_collector.o \
 	lexer.o \
 	match_identifiers.o \
 	parser.o \
@@ -23,8 +18,13 @@ OBJNAMES := \
 	typechecker.o \
 	typed_ast.o \
 	typesystem.o \
-	value.o \
-	native.o
+	interpreter/environment.o \
+	interpreter/error.o \
+	interpreter/eval.o \
+	interpreter/execute.o \
+	interpreter/garbage_collector.o \
+	interpreter/value.o \
+	interpreter/native.o
 
 OBJS := $(addprefix $(OBJDIR)/,$(OBJNAMES))
 
@@ -41,21 +41,22 @@ clean:
 	rm $(BINDIR) -r
 	rm $(OBJDIR) -r
 
-$(BINDIR)/jasperi: $(OBJS) $(OBJDIR)/main.o | $(BINDIR)
+$(BINDIR)/jasperi: $(OBJS) $(OBJDIR)/interpreter/main.o | $(BINDIR)
 	$(CXX) -o $@ $^ $(LIBS)
 
 $(BINDIR)/test_program: $(OBJS) $(OBJDIR)/tests.o | $(BINDIR)
 	$(CXX) -o $@ $^ $(LIBS)
 
+$(OBJS) $(OBJDIR)/interpreter/main.o $(OBJDIR)/tests.o: | $(OBJDIR)
+
 $(BINDIR):
 	mkdir $(BINDIR)
 
-$(OBJS) $(OBJDIR)/main.o $(OBJDIR)/tests.o: | $(OBJDIR)
-
 $(OBJDIR):
 	mkdir $(OBJDIR)
+	mkdir $(OBJDIR)/interpreter
 
-include $(OBJDIR)/main.d
+include $(OBJDIR)/interpreter/main.d
 include $(OBJDIR)/tests.d
 include $(OBJS:.o=.d)
