+++ src/interpreter/garbage_collector.cpp
