+++ src/test_set.hpp
@@ -3,8 +3,8 @@
 #include <vector>
 #include <string>
 
-#include "exit_status_type.hpp"
-#include "environment_fwd.hpp"
+#include "interpreter/exit_status_type.hpp"
+#include "interpreter/environment_fwd.hpp"
 
 namespace Test {
